[analyzer] Move enum name error logic out of MemberDuplicateDefinitionVerifier.
Moves the logic for reporting `diag.enumConstantSameNameAsEnclosing`
and `diag.enumWithNameValues` out of
`MemberDuplicateDefinitionVerifier` and into `ErrorVerifier`. The
logic for detecting these diagnostics doesn't really make sense to be
part of `MemberDuplicateDefinitionVerifier`, because
`MemberDuplicateDefinitionVerifier` is concerned with detecting when a
class member has the same name as a previous member, whereas
`diag.enumConstantSameNameAsEnclosing` and `diag.enumWithNameValues`
are concerned with detecting when a class member has the same as the
class itself.

In a follow-up CL I intend to do a pretty substantial rewrite of
`MemberDuplicateDefinitionVerifier`. Moving this unrelated logic out
of `MemberDuplicateDefinitionVerifier` will make that rewrite more
straightforward.

Change-Id: I6a6a69645b1b3e5c4524d6de5816cb687b14994b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/468841
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>