[analyzer] Add ClassMemberImpl.augmentKeyword.
All classes derived from `ClassMemberImpl` already had an
implementation of `augmentKeyword` except for
`PrimaryConstructorBodyImpl` (which was added as part of this CL).

In a follow-up CL, this will be used as part of a rewrite of
`MemberDuplicateDefinitionVerifier`, to allow members marked `augment`
to be excluded from duplicate checking.

Change-Id: I6a6a6964b8e41e3e5fde3e17428204d1e0898bbf
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/468882
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>