[cfe][PrimaryConstructors] Delay adding initializers
This simplifies how constructor initializers are added to the SourceConstructorBuilder. Previously the initializers were added from both within the body builder, prior to inference, and in the resolver, post inference; and initializer consistency was checked for each addition.

Now all initializers are collected during body building, inferred, and then checked for consistency before being added to the SourceConstructorBuilder.

This change paves the way for handling primary constructor body declarations which requires a more high-level handling due to it splitting the definition of primary constructors in the two distinct syntactical units which needs to be handled partially, one at a time, during body building.

Change-Id: I56d376022dfe77340a4a5beee074610d1e3119c6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/468780
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>