[messages] Fix context message modification bug.
Previously, if `DiagnosticWithoutArguments.withContextMessages` was
used to attach context messages to a diagnostic code that didn't
require arguments, the list of context messages supplied by the error
reporting logic would be stored directly in the returned
`LocatableDiagnostic` object. If
`DiagnosticReporter._createDiagnostic` later tried to add more context
messages in order to disambiguate types, it would try to add them
directly to that list.

This was problematic for two reasons:

- The list might have been a const list, or a list whose runtime type
  was a proper subtype of `List<DiagnosticMessage>` (for example
  `List<DiagnosticMessageImpl>`), in which case attempting to add more
  context messages would result in a runtime exception.

- There was no guarantee that the `LocatableDiagnostic` object had an
  exclusive pointer to the list; the list might have been re-used by
  the error reporting logic for other diagnostics. If this happened,
  then all references to the list would receive the additional context
  messages.

To fix this, I've changed
`DiagnosticWithoutArguments.withContextMessages` so that it makes a
copy of the incoming list. This guarantees that the list in the
returned `LocatableDiagnostic` object is always precisely of type
`List<DiagnosticMessage>`, is modifiable, and is not pointed to by any
other object.

The additional copying should have a negligible impact on performance,
since this code path is only exercised when there are diagnostics to
report.

I've also changed the signature of
`LocatableDiagnostic.withContextMessages` so that it accepts
`Iterable<DiagnosticMessage>` rather than `List<DiagnosticMessage>`,
since all implementations of this method now make a copy.

I discovered this issue while preparing a follow-up CL that rewrites
the `MemberDuplicateDefinitionVerifier` class; the new version of the
class happens to exercise the problematic code path.

Change-Id: I6a6a6964ac8fdbfab4517345b17502a0f3d50c55
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/468906
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>